wip(#1102): extract emit/stmt.rs (emit_stmt + expr_has_value + has_return)
